Add basic error page, work towards URL handling functions
Now, URLs are handled by functions. Basically, you have a regex and a function
which takes a string and returns an Element (the page body). If the URL given
to DGI matches the regex, that function is called with the URL requested to
obtain what we should put out.

In the event that there are any exceptions, we just bail out and give a basic
error page which is a text import.