Handle displaying posts more completely, lots of error cases covered
If there are no updates at all, let the reader know. If updates don't go as
high as they want, let them know. If we can't parse the post because jac is a
moron, let the useer know ;) If the post is good, display it properly. If the
post is requested by title, give a notice of that but don't yet do the proper
thing.