Split off commit information generation into it's own function
The thing after the repo name in the URL is now a command. If it is "commits"
we show a full listing of commits (with their subjects chopped to
MAX_SUBJECT_LENGTH if need be). We also use the same function to display the
most recent 10 commits on the general page.