Alter .png, .css, and .js handling to be automatic
Instead of looking for specific files or in specific directories, we do a
blanket "Does the requested file exist?" If it does, we check to see if it's
one of the approved types. We output the correct content type header, and then
dump up to the first 128KiB of the file.