Alter repo table to use our own tab setup
There aren't many changes. The main id was switched to the class tabbox, and
instead of having to make our own list of tabs, it is generated for us based
on the title we give each tab. Also, each tab just has the class "tab" instead
of an id increasing in number.