Gzip things ourself, add max-age to headers
This commit should really have been two. The first alterst the signature of
the handlers to accept a ref string to put headers in. We use this to be able
to tell ourselves what the max-age should be for content expiry. It defaults
to 60 seconds if we don't give any headers.

The second commit is more influential in that it adds gzip support through the
zlib library. What this means, is we now have a function we call to write the
document and headers at the same time. This checks to see if the browser
supports gzip and if it does, we send the gzipped version of everything
instead.

There is only one place this is not true, where if a .gz exists on the system
we take care of sending it ourself (and avoid re-gzipping it). However, this
does not check to see if the client supports it or not.