Rework Makefile entirel, include js. See full log
The major work present here is the complete reworking of the make system.
Instead of having objects in src/, they go to obj/. Instead of having four
extra scripts to deploy the binary as a tar or patch the remote binary, we now
have a number of targets that culminate with the targets deploy and cdeploy.
deploy attempts to patch the remote binary. cdeploy compresses the binary and
copies over a brand new one.