Update old code, add file listing handling code, see full log
getPath is no longer just a dummy function. The file listing handler is fairly
simple. It simply uses api/cat to get the contents of the current branch/path
and displays it. If the current path is a file, it dumps the file instead.
Above the file listing/file, is a div with the current (clickable) path.